pyproject.toml
[tool.poetry]
name = "FastAPI_Pg_Counter_App"
version = "1.0.0"
description = "Counter_app FastAPI and Postgres interaction"
authors = ["Dmitry Ilyushko <dm.ilyushko@gmail.com>"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pre-commit]
hooks = [
    "black",
    "flake8",
    "isort",
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.27.0"
fastapi = "^0.109.0"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
jinja2 = "^3.1.3"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 
